soc-threat-intel-assistant/requirements.txt
rich
requests
python-dotenv
python-whois
whois
